feat: 라우팅 연동 및 텍스트 위 드래그 방지
