+++ src/components/app/home/recent-card/link-and-go-button.tsx
